requirements.txt
python-dotenv
openai
numpy
pandas
openpyxl
flask